+++ tests/test_subscriber.py
@@ -1,15 +1,45 @@
 from subscriber import podaac_data_subscriber as pds
+from subscriber import podaac_access as pa
 import pytest
-
+import os
+from pathlib import Path
+import shutil
 
 def test_temporal_range():
 
-    assert pds.get_temporal_range(None, '2021-01-01T00:00:00Z', "2021-08-20T13:30:38Z") == "1900-01-01T00:00:00Z,2021-01-01T00:00:00Z"
-    assert pds.get_temporal_range('2021-01-01T00:00:00Z', '2022-01-01T00:00:00Z', "2021-08-20T13:30:38Z") == "2021-01-01T00:00:00Z,2022-01-01T00:00:00Z"
-    assert pds.get_temporal_range('2021-01-01T00:00:00Z', None, "2021-08-20T13:30:38Z") == "2021-01-01T00:00:00Z,2021-08-20T13:30:38Z"
+    assert pa.get_temporal_range(None, '2021-01-01T00:00:00Z', "2021-08-20T13:30:38Z") == "1900-01-01T00:00:00Z,2021-01-01T00:00:00Z"
+    assert pa.get_temporal_range('2021-01-01T00:00:00Z', '2022-01-01T00:00:00Z', "2021-08-20T13:30:38Z") == "2021-01-01T00:00:00Z,2022-01-01T00:00:00Z"
+    assert pa.get_temporal_range('2021-01-01T00:00:00Z', None, "2021-08-20T13:30:38Z") == "2021-01-01T00:00:00Z,2021-08-20T13:30:38Z"
     with pytest.raises(ValueError):
-        pds.get_temporal_range(None, None, None) == "2021-01-01T00:00:00Z,2021-08-20T13:30:38Z"
+        pa.get_temporal_range(None, None, None) == "2021-01-01T00:00:00Z,2021-08-20T13:30:38Z"
 
+
+
+data_dir_with_updates = "./test_update_format_change"
+
+@pytest.fixture
+def cleanup_update_test():
+    yield None
+    print("Cleanup...")
+    shutil.rmtree(data_dir_with_updates)
+
+
+def test_update_format_change(cleanup_update_test):
+    print("Running Test")
+    data_dir_with_updates = "./test_update_format_change"
+    data_dir_with_no_updates = "./test_update_format_change_empty"
+
+    collection_name_does_not_exist = "collectionNameNotfound"
+    collection_name_exists = "collectionName"
+
+    # Make directories and files...
+    os.makedirs(data_dir_with_updates, exist_ok=True )
+    Path(data_dir_with_updates+'/.update').touch()
+    Path(data_dir_with_updates+'/.update__' + collection_name_exists).touch()
+
+    assert pds.get_update_file(data_dir_with_no_updates, collection_name_does_not_exist) == None
+    assert pds.get_update_file(data_dir_with_updates, collection_name_does_not_exist) == data_dir_with_updates+"/.update"
+    assert pds.get_update_file(data_dir_with_updates, collection_name_exists) == data_dir_with_updates+'/.update__' + collection_name_exists
 
 def test_validate():
     # work
@@ -94,5 +124,5 @@ def test_validate():
 def validate(args):
     parser = pds.create_parser()
     args2 = parser.parse_args(args)
-    pds.validate(args2)
+    pa.validate(args2)
     return args2